feat(opts): Set notifications on/off in opts
This commit adds the `hbac_notify` function to `utils` to simplify all notifications generally and to allow the user to suppress notifications if desired.

The notifications are useful for affirming that an Hbac action executed successfully, but a user might not want it or have other visual cues for the pin state of loaded buffers, such as a status/bufferline icon.

docs(readme): Add notify opt to readme