feat: Store current pinned bufs in json file
Adds `storage` module and `store_pinned_bufs` function.
The func checks for an existing `pin_storage.json` file in the nvim datadir (or creates a new one) and updates it with a named entry (by user input, or defaults to the timestamp) of all the currently pinned bufnames, their relative path, the current project root, and the timestamp. If a key already exists with the chosen name, the user will be asked to confirm the overwrite.

This will be used to load previously pinned buffers with a func call or a Telescope picker.