feat(storage): Open previously stored pinned bufs
Adds `open_pin_storage_entry` func. Takes a storage entry keyname as an arg. Warns and early returns if key doesn't exist. Opens each file with an `:e` command and sets the pin state of the new buffer to true.

Adds `abs_path` key to each stored pin, which is the full path used to open a pin. If the file at the path doesn't exist, the `:e` command is ran anyway, on the assumption that a user might want to use the storage entry as a template for some kind of project structure. It may be a good idea to add a warning that these files don't exist.

Adds `storage.open` table to the setup opts table. `storage.open` has two keys: `prehook` and `posthook`, which take a function as a value. They are run before and after the stored pins in an entry are opened. For example, a user might want to open a new tab with `vim.cmd("tabnew")` or close all unpinned bufs with `:Hbac close_unpinned` before opening a stored entry.