feat(storage): Use a Telescope picker to open stored pins
Adds `storage.telescope` and related modules.

`require("hbac.storage.telescope").stored_pins_picker()` will open the picker. Each entry in the picker is the keyname of the stored pins, followed by the number of stored pins in that key, followed by the project root where that key was created. The preview shows the datetime when the key was created followed by a list of the stored pins, using the same filename/filepath format used by the Hbac pin picker.