feat(storage): Customize function for opening stored pin
Clarifies in readme that prehook is called before any stored pin is open and posthook is called after all are opened

The new `on_open` function, which is now customizable in the opts, is the only function called during each iteration through the stored pins entry. By default, it opens the stored pin in the current window and sets its pin state to true in `state.pinned_buffers`, matching the behavior of the corresponding code in prior commits.