fix(storage): Abort storing pins during entry input if blank
Previously, if the entry name was left blank while pins were being stored, the entry name would default to the timestamp. However, if a user wanted to abort during the input prompt by pressing escape, the keyname would still be stored as "", and thus default to the timestamp.

In this commit, we change the behavior of the keyname registering function. If the keyname is left blank, we interpret this as a user's intention to abort. If the user wants to use a timestamp, they can input `%t` for the keyname.