feat(storage): Update existing storage entry with arg to store_pinned_bufs
If no arg is passed to store_pinned_bufs, then the user will be prompted for a new entry name, as in prior commits.
If an arg is passed and it doesn't match an existing entry keyname, then a new entry will be written using the existing pins immediately.
If an arg is passed and it DOES match an existing entry keyname, the user will be prompted to confirm the update of the existing entry, and the entry will be overwritten.

fix(storage): Give proper write-type message on updates