feat(storage-telescope): Update selected entries w/ current pinned bufs
With default key `<M-u>`, opens a prompt to overwrite the selected entries in the storage telescope picker with the currently pinned buffers. The intended way to remove/add single buffers from an entry is to load the storage entry, pin/unpin the desired buffers, and overwrite it.
This was possible with prior features (e.g. by using the 'Hbac store_pinned_bufs' command and entering the name of an existing entry) but the picker lets a viewer view and search for existing entries and updated them without needing to type the entry name

docs(readme): Add Telescope mapping defaults, cleanup examples