feat(storage): Add current buffer to selected entries
`hbac.storage.add_cur_buf_to_entry` takes a keyname of a storage entry as an argument. The file in the current buffer will be added to the entry if it's not already in it.

The file in the current buffer can be added to any entries selected by the storage picker with the `hbac_add_cur_buf_to_entry` action (default `<M-b>`).

refactor(storage): Check if file is in stored pins

Makes the function `pin_is_duplicae_in_entry` more general rather than specifically for warning and preventing duplicate entries

fix(storage): Get cur buf when calling add_cur_buf_to_entry func, not the most recent cur bur when the storage picker was opened