feat(telescope): Remove files from storage with storage_entry_previewer
As in prior commits, the storage entry previewer can be opened from the storage picker with `<C-p>` by default over an entry selection.
This commit adds the `hbac_remove_file_from_entry` action to the storage entry previewer picker (default `<M-x>`). This removes all selected results in the picker from the parent storage entry. The storage file is updated immediately and the previewer picker is refreshed.

fix(telescope): move cursor to bottom instead of up a row when acting on the last row