fix(telescope): Pass opts to make_finder in Telescope pickers
Copies/modifies the code in telescope.builtin.buffers to filter the bufnr array based on user opts.

Adds a most_recent_buf func to utils. We use this instead of vim.api.nvim_get_current_buf because depending on how we're jumping around buffers, we don't want to necessarily get the current buf, since it might be an empty buffer (e.g. a prompt type buf as we're going from picker to picker), but rather the last used listed buffer.
This most_recent_buf func is used to filter out the current buf in the picker if ignore_current_buffer is set in the picker opts table.

The storage picker and its companion file-previewer picker also get an opts table, though not much is being done with them. We're just passing it now in case we implement some opts in the future.