feat(telescope): Add pin sorter command to pin_picker
Using the action `hbac_add_buf_to_entry` (default `<M-b>`) from the pin picker, the selected buffers from the results (regardless of whether they are "pinned" or "unpinned") will be stored in an array. A new Telescope picker, the "pin sorter", will pop up. It is essentially the same picker as the storage picker with different default opts.

The main differences between the pin sorter and the storage picker are the default mappings.

- <CR>
<CR> will add the buffers we just stored into all the selected storage entries. This is done with the renamed `add_buf_to_entry` function (the same one called in the `hbac_add_buf_to_entry` action from the regular storage picker, default `<M-b>`). If the file that belongs to the buffer is already in the entr(ies) that we're trying to add to, a warning notification will pop up, and nothing will be added to the entr(ies). Otherwise, the file info will be added as we would expect from other pin storage functions.

- <Esc>
<Esc> will recall the pin picker without closing it, similar to how we can go back and forth from the storage picker to the storage previewer with <C-p> and <Esc>

Otherwise, there are aesthetic opts differences for the sake of trying to distinguish it from the storage picker, though this may not be the case depending on the user's Telescope defaults.